end to end execution of the pipeline

+++ dbt/urbangear/models/marts/fct_daily_sales.sql
@@ -1,20 +1,52 @@
-{{ config(materialized='incremental', unique_key='order_id', schema='analytics', tags=['daily']) }}
+{{ 
+    config(
+        materialized='incremental', 
+        unique_key='order_id', 
+        schema='analytics', 
+        tags=['daily']
+    ) 
+}}
 
 WITH orders AS (
     SELECT * FROM {{ ref('stg_orders') }}
     {% if is_incremental() %}
-    WHERE order_date::date > (SELECT COALESCE(MAX(order_date::date) - INTERVAL '3 days', '1970-01-01'::DATE) FROM {{ this }})
+    WHERE order_date::date > (
+        SELECT COALESCE(MAX(order_date::date) - INTERVAL '3 days', '1970-01-01'::DATE) 
+        FROM {{ this }}
+    )
     {% endif %}
 )
+
 SELECT
-    order_id, customer_id, platform, order_date, product_id, product_name, category,
-    quantity, unit_price, discount_pct,
-    ROUND(quantity * unit_price,2) AS gross_revenue,
-    ROUND(quantity * unit_price * discount_pct / 100.0,2) AS discount_amount,
-    ROUND(quantity * unit_price * (1 - discount_pct / 100.0),2) AS net_revenue,
-    shipping_cost, tax_amount,
-    ROUND(quantity * unit_price * (1 - discount_pct / 100.0) + shipping_cost + tax_amount,2) AS total_amount,
-    payment_method, shipping_country, shipping_state, shipping_city, is_returned,
+    order_id, 
+    customer_id, 
+    platform, 
+    order_date, 
+    product_id, 
+    product_name, 
+    category,
+    quantity, 
+    unit_price, 
+    discount_pct,
+    -- Revenue measures
+    ROUND( (quantity * unit_price)::numeric, 2) AS gross_revenue,
+    ROUND(  (quantity * unit_price * discount_pct / 100)::numeric, 2) AS discount_amount,
+    ROUND( (quantity * unit_price * (1 - discount_pct / 100))::numeric, 2) AS net_revenue,
+    shipping_cost, 
+    tax_amount,
+    ROUND(
+        (
+            quantity * unit_price * (1 - discount_pct / 100)
+            + shipping_cost + tax_amount
+        )::numeric,
+        2
+    ) AS total_amount,
+    payment_method, 
+    shipping_country, 
+    shipping_state, 
+    shipping_city, 
+    is_returned,
+    -- Convenience date columns
     DATE_TRUNC('day', order_date)::DATE AS order_day,
     DATE_TRUNC('week', order_date)::DATE AS order_week,
     DATE_TRUNC('month', order_date)::DATE AS order_month,
